hotfix: replace gokv with deno cache api
