+++ .gitignore
@@ -14,3 +14,6 @@ Thumbs.db
 !.env.vault
 
 !.env.project
+.flaskenv*
+
+deno.lock
