+++ README.md
@@ -175,4 +175,3 @@ MIT © [Nicholas Berlette][nberlette]
 [example-2]: https://migo.deno.dev/image.png?icon=twitter&subtitleFontSize=48&iconColor=0cf&titleY=460&subtitleFontSize=48&title=Creating%20Dynamic%20Cover%20Images&subtitle=By%20Nicholas%20Berlette&borderRadius=25&titleFontFamily=monospace
 [example-3]: https://migo.deno.dev/image.png?icon=nuxtdotjs&bgColor=112233&iconColor=00DC82&iconStroke=00DC82&iconStrokeWidth=0.55&titleColor=00DC82&subtitleColor=e0e0e0&iconW=300&iconY=50&titleY=460&title=Nuxt%20ContentWind%20Starter&subtitle=stackblitz.com%2fedit%2fcontent-wind&borderRadius=30&pxRatio=1.5
 [example-4]: https://migo.deno.dev/image.png?title=deno911⁄+kit&subtitle=this+one+has+custom+alignment+and+size!&titleFontFamily=sans-serif&titleFontSize=80&titleFontWeight=900&titleTextAnchor=right&titleX=160&titleY=115&subtitleFontSize=36&subtitleFontWeight=900&subtitleFontFamily=monospace&subtitleTextAnchor=left&subtitleX=40&subtitleY=260&pxRatio=1.5&width=1000&height=300&bgColor=fff&titleColor=123&subtitleColor=456&icon=noto:sauropod&iconW=100&iconH=100&iconX=40&iconY=30&borderRadius=20
-
