+++ deps.ts
@@ -1,19 +1,25 @@
 export { formatHex, parse as parseColor } from "culori";
-export { default as GOKV } from "gokv";
-export { html } from "htm";
-export { UnoCSS } from "htm/plugins.ts";
+export { default as html } from "htm/mod.ts";
+export { default as UnoCSS } from "htm/plugins/unocss.ts";
+export { default as ColorScheme } from "htm/plugins/color-scheme.ts";
+export * from "htm/mod.ts";
+
 export { default as presetWind } from "@unocss/preset-wind.ts";
 export { type ConnInfo, type Routes, serve, Status, STATUS_TEXT } from "sift";
-export * from "preact";
+// export * from "preact";
 export { render as renderToString } from "preact/render";
 export { default as colorHash } from "colorhash";
 export { render as rasterizeSVG } from "resvg";
+export * as etag from "etag";
+export * from "is";
+export * from "serialize";
+import { Logger } from "911";
+export const log = new Logger();
 
 export {
   camelCase,
   decode,
   encode,
-  eTag as etag,
   groupBy,
   isArray,
   isObject,
@@ -31,6 +37,3 @@ export {
   utf8TextDecoder,
   utf8TextEncoder,
 } from "911";
-export { default as is } from "is";
-import { Logger } from "911";
-export const log = new Logger();
