+++ import_map.json
@@ -4,33 +4,38 @@
     "/": "./",
     "./": "./",
     "src/": "./src/",
-    "lib/": "./lib/",
     "deps.ts": "./deps.ts",
     "~/deps": "./deps.ts",
     "~/deps.ts": "./deps.ts",
-    "std/": "https://deno.land/std@0.155.0/",
+    "#imports": "./deps.ts",
+    "#utils": "./src/utils.ts",
+    "#constants": "./src/constants.ts",
+    "std/": "https://deno.land/std@0.166.0/",
     "esm/": "https://esm.sh/",
     "preact": "https://esm.sh/preact@10.10.6",
     "preact/render": "https://esm.sh/preact-render-to-string@5.2.1?external=preact&target=deno",
     "preact/jsx-runtime": "https://esm.sh/preact@10.10.6/jsx-runtime",
     "preact/jsx-dev-runtime": "https://esm.sh/preact@10.10.6/jsx-dev-runtime",
     "preact/": "https://esm.sh/preact@10.10.6/",
-    "is": "https://deno.land/x/dis@0.0.1/mod.ts",
     "case": "https://deno.land/x/case@2.1.1/mod.ts",
     "culori": "https://deno.land/x/culori@v2.1.0-alpha.0/index.js?dts",
     "culori/": "https://deno.land/x/culori@v2.1.0-alpha.0/",
-    "gokv": "https://deno.land/x/gokv@0.0.12/mod.ts",
-    "htm": "https://deno.land/x/htm@0.0.10/mod.tsx",
-    "htm/": "https://deno.land/x/htm@0.0.10/",
-    "sift": "https://denopkg.com/nberlette/sift@main/mod.ts",
-    "sift/": "https://denopkg.com/nberlette/sift@main/",
+    "htm": "https://deno.land/x/htm@0.1.2/mod.ts",
+    "htm/": "https://deno.land/x/htm@0.1.2/",
+    "sift": "https://deno.land/x/sift@0.6.0/mod.ts",
+    "sift/": "https://deno.land/x/sift@0.6.0/",
     "colorhash": "https://deno.land/x/color_hash@v2.0.1/mod.ts",
     "resvg": "https://deno.land/x/resvg_wasm@0.2.0/mod.ts",
     "aleph/server": "https://deno.land/x/aleph/server/mod.ts",
     "aleph/": "https://deno.land/x/aleph@1.0.0-beta.1/",
     "@unocss/": "https://deno.land/x/aleph@1.0.0-beta.1/lib/@unocss/",
     "911": "https://deno.land/x/911@0.1.5/mod.ts",
-    "911/": "https://deno.land/x/911@0.1.5/src/"
+    "911/": "https://deno.land/x/911@0.1.5/src/",
+    "etag": "https://deno.land/x/etag@0.0.1/mod.ts",
+    "is": "https://deno.land/x/dis@0.2.0/mod.ts",
+    "this": "https://deno.land/x/this@0.160.1/mod.ts",
+    "this/": "https://deno.land/x/this@0.160.1/",
+    "serialize": "https://deno.land/x/serialize@1.0.0-rc.2/mod.ts"
   },
   "scopes": {}
 }
